refactor: move timeEpoch/timestamp into main package files
