+++ browser/chromium/chromium.go
@@ -3,6 +3,7 @@ package chromium
 import (
 	"os"
 	"path/filepath"
+	"time"
 
 	"github.com/moond4rk/hackbrowserdata/crypto/keyretriever"
 	"github.com/moond4rk/hackbrowserdata/filemanager"
@@ -272,4 +273,19 @@ func isSkippedDir(name string) bool {
 		return true
 	}
 	return false
+}
+
+// timeEpoch converts a WebKit/Chromium epoch timestamp (microseconds since
+// 1601-01-01) to a time.Time.
+func timeEpoch(epoch int64) time.Time {
+	maxTime := int64(99633311740000000)
+	if epoch > maxTime {
+		return time.Date(2049, 1, 1, 1, 1, 1, 1, time.Local)
+	}
+	t := time.Date(1601, 1, 1, 0, 0, 0, 0, time.Local)
+	d := time.Duration(epoch)
+	for i := 0; i < 1000; i++ {
+		t = t.Add(d)
+	}
+	return t
 }
